+++ src/dec_borderheur.c
@@ -577,7 +577,7 @@ static SCIP_RETCODE buildTransformedProblem(
    /* go through all of the constraints */
    for( i = 0; i < nconss; i++ )
    {
-      long int consblock = -1;
+      int consblock = -1;
       int ncurvars;
       SCIP_VAR **curvars;
 
@@ -596,7 +596,7 @@ static SCIP_RETCODE buildTransformedProblem(
       for( j = 0; j < ncurvars; j++ )
       {
          SCIP_VAR* var;
-         long int varblock;
+         int varblock;
          assert(curvars != NULL);
 
          if( !SCIPisVarRelevant(curvars[j]) )
@@ -635,7 +635,7 @@ static SCIP_RETCODE buildTransformedProblem(
          }
          else
          {
-            varblock = (long int)SCIPhashmapGetImage(vartoblock, var);
+            varblock = (int)(size_t)SCIPhashmapGetImage(vartoblock, var);
             assert(varblock == detectordata->varpart[SCIPvarGetProbindex(var)] ||  detectordata->varpart[SCIPvarGetProbindex(var)] == -2);
          }
